moving ops to separate file
