+++ pennylane/beta/plugins/default_tensor_tf.py
@@ -30,159 +30,10 @@ except ImportError as e:
 
 from pennylane.variable import Variable
 from pennylane.beta.plugins.default_tensor import DefaultTensor
-from pennylane.beta.plugins.numpy_ops import I, X, Y, Z
-
+from pennylane.beta.plugins import tf_ops as ops
 
 # tolerance for numerical errors
 tolerance = 1e-10
-C_DTYPE = tf.complex128
-R_DTYPE = tf.float64
-
-
-I = tf.constant(I, dtype=C_DTYPE)
-X = tf.constant(X, dtype=C_DTYPE)
-
-II = tf.eye(4, dtype=C_DTYPE)
-ZZ = tf.constant(np.kron(Z, Z), dtype=C_DTYPE)
-
-IX = tf.constant(np.kron(I, X), dtype=C_DTYPE)
-IY = tf.constant(np.kron(I, Y), dtype=C_DTYPE)
-IZ = tf.constant(np.kron(I, Z), dtype=C_DTYPE)
-
-ZI = tf.constant(np.kron(Z, I), dtype=C_DTYPE)
-ZX = tf.constant(np.kron(Z, X), dtype=C_DTYPE)
-ZY = tf.constant(np.kron(Z, Y), dtype=C_DTYPE)
-
-
-def Rphi(phi):
-    r"""One-qubit phase shift.
-
-    Args:
-        phi (float): phase shift angle
-
-    Returns:
-        array: unitary 2x2 phase shift matrix
-    """
-    phi = tf.cast(phi, dtype=C_DTYPE)
-    return ((1 + tf.exp(1j * phi)) * I + (1 - tf.exp(1j * phi)) * Z) / 2
-
-
-def Rotx(theta):
-    r"""One-qubit rotation about the x axis.
-
-    Args:
-        theta (float): rotation angle
-
-    Returns:
-        array: unitary 2x2 rotation matrix :math:`e^{-i \sigma_x \theta/2}`
-    """
-    theta = tf.cast(theta, dtype=C_DTYPE)
-    return tf.cos(theta / 2) * I + 1j * tf.sin(-theta / 2) * X
-
-
-def Roty(theta):
-    r"""One-qubit rotation about the y axis.
-
-    Args:
-        theta (float): rotation angle
-
-    Returns:
-        array: unitary 2x2 rotation matrix :math:`e^{-i \sigma_y \theta/2}`
-    """
-    theta = tf.cast(theta, dtype=C_DTYPE)
-    return tf.cos(theta / 2) * I + 1j * tf.sin(-theta / 2) * Y
-
-
-def Rotz(theta):
-    r"""One-qubit rotation about the z axis.
-
-    Args:
-        theta (float): rotation angle
-
-    Returns:
-        array: unitary 2x2 rotation matrix :math:`e^{-i \sigma_z \theta/2}`
-    """
-    theta = tf.cast(theta, dtype=C_DTYPE)
-    return tf.cos(theta / 2) * I + 1j * tf.sin(-theta / 2) * Z
-
-
-def Rot3(a, b, c):
-    r"""Arbitrary one-qubit rotation using three Euler angles.
-
-    Args:
-        a,b,c (float): rotation angles
-
-    Returns:
-        array: unitary 2x2 rotation matrix ``rz(c) @ ry(b) @ rz(a)``
-    """
-    return Rotz(c) @ Roty(b) @ Rotz(a)
-
-
-def CRotx(theta):
-    r"""Two-qubit controlled rotation about the x axis.
-
-    Args:
-        theta (float): rotation angle
-
-    Returns:
-        array: unitary 4x4 rotation matrix
-        :math:`|0\rangle\langle 0|\otimes \mathbb{I}+|1\rangle\langle 1|\otimes R_x(\theta)`
-    """
-    theta = tf.cast(theta, dtype=C_DTYPE)
-    return (
-        tf.cos(theta / 4) ** 2 * II
-        - 1j * tf.sin(theta / 2) / 2 * IX
-        + tf.sin(theta / 4) ** 2 * ZI
-        + 1j * tf.sin(theta / 2) / 2 * ZX
-    )
-
-
-def CRoty(theta):
-    r"""Two-qubit controlled rotation about the y axis.
-
-    Args:
-        theta (float): rotation angle
-    Returns:
-        array: unitary 4x4 rotation matrix :math:`|0\rangle\langle 0|\otimes \mathbb{I}+|1\rangle\langle 1|\otimes R_y(\theta)`
-    """
-    theta = tf.cast(theta, dtype=C_DTYPE)
-    return (
-        tf.cos(theta / 4) ** 2 * II
-        - 1j * tf.sin(theta / 2) / 2 * IY
-        + tf.sin(theta / 4) ** 2 * ZI
-        + 1j * tf.sin(theta / 2) / 2 * ZY
-    )
-
-
-def CRotz(theta):
-    r"""Two-qubit controlled rotation about the z axis.
-
-    Args:
-        theta (float): rotation angle
-    Returns:
-        array: unitary 4x4 rotation matrix
-        :math:`|0\rangle\langle 0|\otimes \mathbb{I}+|1\rangle\langle 1|\otimes R_z(\theta)`
-    """
-    theta = tf.cast(theta, dtype=C_DTYPE)
-    return (
-        tf.cos(theta / 4) ** 2 * II
-        - 1j * tf.sin(theta / 2) / 2 * IZ
-        + tf.sin(theta / 4) ** 2 * ZI
-        + 1j * tf.sin(theta / 2) / 2 * ZZ
-    )
-
-
-def CRot3(a, b, c):
-    r"""Arbitrary two-qubit controlled rotation using three Euler angles.
-
-    Args:
-        a,b,c (float): rotation angles
-    Returns:
-        array: unitary 4x4 rotation matrix
-        :math:`|0\rangle\langle 0|\otimes \mathbb{I}+|1\rangle\langle 1|\otimes R(a,b,c)`
-    """
-    return CRotz(c) @ (CRoty(b) @ CRotz(a))
-
 
 class DefaultTensorTF(DefaultTensor):
     """Experimental TensorFlow Tensor Network simulator device for PennyLane.
@@ -233,15 +84,15 @@ class DefaultTensorTF(DefaultTensor):
     _operation_map = copy.copy(DefaultTensor._operation_map)
     _operation_map.update(
         {
-            "PhaseShift": Rphi,
-            "RX": Rotx,
-            "RY": Roty,
-            "RZ": Rotz,
-            "Rot": Rot3,
-            "CRX": CRotx,
-            "CRY": CRoty,
-            "CRZ": CRotz,
-            "CRot": CRot3,
+            "PhaseShift": ops.Rphi,
+            "RX": ops.Rotx,
+            "RY": ops.Roty,
+            "RZ": ops.Rotz,
+            "Rot": ops.Rot3,
+            "CRX": ops.CRotx,
+            "CRY": ops.CRoty,
+            "CRZ": ops.CRotz,
+            "CRot": ops.CRot3,
         }
     )
 
@@ -255,8 +106,8 @@ class DefaultTensorTF(DefaultTensor):
     _squeeze = staticmethod(tf.squeeze)
     _expand_dims = staticmethod(tf.expand_dims)
 
-    C_DTYPE = C_DTYPE
-    R_DTYPE = R_DTYPE
+    C_DTYPE = ops.C_DTYPE
+    R_DTYPE = ops.R_DTYPE
 
     def __init__(self, wires, shots=1000, representation="exact", contraction_method="auto"):
         self.variables = []
